chore: fix lint and merge main for worker runtime PR
Merge origin/main and split oversized worker-editable-values tests;
extract editable worker draft handlers to satisfy Biome limits.
